Use patched ansible provisioner for Windows host
On Windows host, ansible_local provisioner doesn't work with a path
expanding problem discussed at
https://github.com/mitchellh/vagrant/issues/6740

This fixes the issue by loading patched ansible provisioner in a way
described at http://www.ryuzee.com/contents/blog/6657

The patch is based on https://github.com/mitchellh/vagrant/pull/6800
It also fixes the Ansible 2.0 integration problem so the shell
provisioner is no longer needed. It is removed.

This is a workaround for Vagrant 1.8.1 or lower so when Vagrant 1.8.2 is
released, this should be removed.