Fix App name and nginx config
The application names were defined as 'bitkids-api' and 'bitkids-ws'
temporally but these were renamed and merged into 'bitkidsserver'. This
fixes the application names in provisioning scripts.

This also changed nginx configuration. It used to redirect '/api/' to
the application API. This changes it to redirect simple '/' access to
the application API.