Add and fix Puma unix domain socket path
nginx redirects a site access to the Rails application via an unix
domain socket, not a TCP connection for performance but Puma default
configuration doesn't have such a listener. This adds it.

Usually, an application specific unix domains socket file is placed in
an application's tmp/sockets directory but VirtualBox vboxsf doesn't
allow to make an unix domain socket file in its filesystem. So, instead
of there, it put the application's unix domain socket file into
/tmp/sockets.